requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.5
discord.py==1.7.3
frozenlist==1.3.1
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
Random-Word==1.0.10
requests==2.28.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
yarl==1.8.1
